requirements.txt
pyyaml==6.0.1
requests==2.32.3
pyvmomi==8.0.2.0.1